requirements.txt
babel==2.17.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.3.0
pillow==11.2.1
pyodbc==5.2.0
tkcalendar==1.6.1
Werkzeug==3.1.4
